refactor(tui): carve chat.rs into tui/chat module
Split the chat TUI implementation into tui/chat modules, carry forward the theme-token base changes, and harden the secret guard against benign documented repo/worktree path false positives while preserving high-entropy detection.

Verification:
- GitHub CI Rust checks passed
- GitHub CI Secret guard passed
- Local cargo fmt --check
- Local cargo clippy --workspace --all-targets -- -D warnings
- Local cargo test --workspace --locked
- Local python3 scripts/check-secrets-test.py
- Local python3 scripts/check-secrets.py
