+++ crates/coven-cli/src/tui/chat/app.rs
@@ -1,0 +1,545 @@
+//! Chat application state, behavior, and helpers. Owns `App` and all of its
+//! methods; provides `discover_agents` and the spinner-frame data.
+
+use std::time::{Duration, Instant};
+
+use crate::harness;
+
+// ── Data types ─────────────────────────────────────────────────────────────
+
+#[derive(Clone, Debug)]
+pub enum MessageRole {
+    User,
+    Agent,
+    System,
+}
+
+#[derive(Clone, Debug)]
+pub struct ChatMessage {
+    pub role: MessageRole,
+    pub sender: String,
+    pub content: String,
+    pub timestamp: String,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub struct AgentInfo {
+    pub id: String,
+    pub label: String,
+    pub harness: String,
+    pub available: bool,
+}
+
+#[derive(Clone, Debug, PartialEq, Eq)]
+pub(super) enum InputMode {
+    Normal,
+    AgentSelect,
+}
+
+#[derive(Clone, Debug)]
+pub(super) enum SlashCommandResult {
+    Handled,
+    Quit,
+    #[allow(dead_code)]
+    Unknown(String),
+}
+
+// ── App state ──────────────────────────────────────────────────────────────
+
+pub(super) struct App {
+    pub(super) messages: Vec<ChatMessage>,
+    pub(super) input: String,
+    pub(super) cursor_pos: usize,
+    pub(super) scroll_offset: usize,
+    pub(super) agents: Vec<AgentInfo>,
+    pub(super) active_agent: Option<usize>,
+    pub(super) input_mode: InputMode,
+    pub(super) agent_select_index: usize,
+    pub(super) show_help: bool,
+    pub(super) spinner_frame: usize,
+    pub(super) is_responding: bool,
+    pub(super) last_tick: Instant,
+}
+
+pub(super) const SPINNER_FRAMES: &[&str] = &["", "", "", "", "", "", "", ""];
+
+impl App {
+    pub(super) fn new() -> Self {
+        let agents = discover_agents();
+        let active_agent = agents.iter().position(|a| a.available);
+
+        let mut app = App {
+            messages: Vec::new(),
+            input: String::new(),
+            cursor_pos: 0,
+            scroll_offset: 0,
+            agents,
+            active_agent,
+            input_mode: InputMode::Normal,
+            agent_select_index: 0,
+            show_help: false,
+            spinner_frame: 0,
+            is_responding: false,
+            last_tick: Instant::now(),
+        };
+
+        app.push_system_message(
+            "Welcome to the Coven. Type a message to chat, or /help for commands.",
+        );
+
+        if let Some(idx) = app.active_agent {
+            let agent = &app.agents[idx];
+            app.push_system_message(&format!(
+                "Active agent: {} ({})",
+                agent.label, agent.harness
+            ));
+        } else {
+            app.push_system_message("No agents available. Run `coven doctor` to check your setup.");
+        }
+
+        app
+    }
+
+    pub(super) fn push_system_message(&mut self, content: &str) {
+        self.messages.push(ChatMessage {
+            role: MessageRole::System,
+            sender: "coven".into(),
+            content: content.to_string(),
+            timestamp: timestamp_now(),
+        });
+    }
+
+    fn push_user_message(&mut self, content: &str) {
+        self.messages.push(ChatMessage {
+            role: MessageRole::User,
+            sender: "You".into(),
+            content: content.to_string(),
+            timestamp: timestamp_now(),
+        });
+    }
+
+    fn push_agent_message(&mut self, agent_name: &str, content: &str) {
+        self.messages.push(ChatMessage {
+            role: MessageRole::Agent,
+            sender: agent_name.to_string(),
+            content: content.to_string(),
+            timestamp: timestamp_now(),
+        });
+    }
+
+    pub(super) fn active_agent_label(&self) -> &str {
+        self.active_agent
+            .and_then(|idx| self.agents.get(idx))
+            .map(|a| a.label.as_str())
+            .unwrap_or("none")
+    }
+
+    pub(super) fn active_agent_harness(&self) -> &str {
+        self.active_agent
+            .and_then(|idx| self.agents.get(idx))
+            .map(|a| a.harness.as_str())
+            .unwrap_or("—")
+    }
+
+    pub(super) fn handle_input(&mut self) -> Option<SlashCommandResult> {
+        let raw = self.input.trim().to_string();
+        self.input.clear();
+        self.cursor_pos = 0;
+
+        if raw.is_empty() {
+            return Some(SlashCommandResult::Handled);
+        }
+
+        if raw.starts_with('/') {
+            return Some(self.handle_slash_command(&raw));
+        }
+
+        // Regular chat message
+        self.push_user_message(&raw);
+        self.simulate_agent_response(&raw);
+        self.scroll_to_bottom();
+        Some(SlashCommandResult::Handled)
+    }
+
+    pub(super) fn handle_slash_command(&mut self, input: &str) -> SlashCommandResult {
+        let parts: Vec<&str> = input.splitn(2, char::is_whitespace).collect();
+        let cmd = parts[0].to_lowercase();
+        let arg = parts.get(1).map(|s| s.trim()).unwrap_or("");
+
+        match cmd.as_str() {
+            "/help" | "/h" => {
+                self.show_help = !self.show_help;
+                SlashCommandResult::Handled
+            }
+            "/clear" | "/cls" => {
+                self.messages.clear();
+                self.scroll_offset = 0;
+                self.push_system_message("Chat cleared.");
+                SlashCommandResult::Handled
+            }
+            "/agent" | "/a" => {
+                if arg.is_empty() {
+                    self.input_mode = InputMode::AgentSelect;
+                    self.agent_select_index = self.active_agent.unwrap_or(0);
+                } else {
+                    self.switch_agent_by_name(arg);
+                }
+                SlashCommandResult::Handled
+            }
+            "/exit" | "/quit" | "/q" => SlashCommandResult::Quit,
+            "/session" | "/sessions" => {
+                self.push_system_message(
+                    "Session management coming soon. Use `coven sessions` in another terminal.",
+                );
+                SlashCommandResult::Handled
+            }
+            "/attach" => {
+                if arg.is_empty() {
+                    self.push_system_message("Usage: /attach <session-id>");
+                } else {
+                    self.push_system_message(&format!(
+                        "Attaching to session {arg}... (coming soon)"
+                    ));
+                }
+                SlashCommandResult::Handled
+            }
+            "/export" => {
+                self.export_chat();
+                SlashCommandResult::Handled
+            }
+            "/run" => {
+                if arg.is_empty() {
+                    self.push_system_message("Usage: /run <harness> <prompt>");
+                } else {
+                    self.push_system_message(&format!("Running: {arg} (coming soon)"));
+                }
+                SlashCommandResult::Handled
+            }
+            "/delegate" => {
+                if arg.is_empty() {
+                    self.push_system_message("Usage: /delegate <agent> <task>");
+                } else {
+                    self.push_system_message(&format!("Delegating: {arg} (coming soon)"));
+                }
+                SlashCommandResult::Handled
+            }
+            "/trace" => {
+                self.push_system_message("Trace display coming soon.");
+                SlashCommandResult::Handled
+            }
+            "/mem" => {
+                if arg.is_empty() {
+                    self.push_system_message("Usage: /mem <query>");
+                } else {
+                    self.push_system_message(&format!(
+                        "Searching agent memory for \"{arg}\"... (coming soon)"
+                    ));
+                }
+                SlashCommandResult::Handled
+            }
+            "/debug" => {
+                self.push_system_message("Debug mode coming soon.");
+                SlashCommandResult::Handled
+            }
+            _ => SlashCommandResult::Unknown(cmd),
+        }
+    }
+
+    pub(super) fn switch_agent_by_name(&mut self, name: &str) {
+        let name_lower = name.to_lowercase();
+        if let Some(idx) = self
+            .agents
+            .iter()
+            .position(|a| a.id.to_lowercase() == name_lower || a.label.to_lowercase() == name_lower)
+        {
+            let agent = &self.agents[idx];
+            if agent.available {
+                self.active_agent = Some(idx);
+                self.push_system_message(&format!(
+                    "Switched to {} ({})",
+                    agent.label, agent.harness
+                ));
+            } else {
+                self.push_system_message(&format!(
+                    "{} is not available. Run `coven doctor` to troubleshoot.",
+                    agent.label
+                ));
+            }
+        } else {
+            let available: Vec<&str> = self.agents.iter().map(|a| a.id.as_str()).collect();
+            self.push_system_message(&format!(
+                "Unknown agent \"{name}\". Available: {}",
+                available.join(", ")
+            ));
+        }
+    }
+
+    pub(super) fn switch_agent_by_index(&mut self, idx: usize) {
+        if let Some(agent) = self.agents.get(idx) {
+            if agent.available {
+                self.active_agent = Some(idx);
+                self.push_system_message(&format!(
+                    "Switched to {} ({})",
+                    agent.label, agent.harness
+                ));
+            } else {
+                self.push_system_message(&format!(
+                    "{} is not available. Run `coven doctor` to troubleshoot.",
+                    agent.label
+                ));
+            }
+        }
+        self.input_mode = InputMode::Normal;
+    }
+
+    fn simulate_agent_response(&mut self, user_msg: &str) {
+        // MVP: show a placeholder response. Real streaming comes in v0.2.
+        let agent_name = self.active_agent_label().to_string();
+        if self.active_agent.is_none() {
+            self.push_system_message(
+                "No active agent. Use /agent to select one, or run `coven doctor`.",
+            );
+            return;
+        }
+
+        self.push_agent_message(
+            &agent_name,
+            &format!(
+                "I received your message: \"{}\"\n\n\
+                 (This is a placeholder response. Real agent streaming will connect \
+                 to the Coven daemon via the session API in v0.2.)\n\n\
+                 To actually run a task, use:\n  \
+                 coven run {} \"{}\"",
+                truncate_str(user_msg, 80),
+                self.active_agent_harness(),
+                truncate_str(user_msg, 60),
+            ),
+        );
+    }
+
+    fn export_chat(&mut self) {
+        if self.messages.is_empty() {
+            self.push_system_message("Nothing to export.");
+            return;
+        }
+
+        let home = dirs_next::home_dir().unwrap_or_default();
+        let export_dir = home.join(".coven").join("exports");
+        if std::fs::create_dir_all(&export_dir).is_err() {
+            self.push_system_message("Failed to create export directory.");
+            return;
+        }
+
+        let filename = format!("chat-{}.md", chrono::Utc::now().format("%Y%m%d-%H%M%S"));
+        let path = export_dir.join(&filename);
+
+        let mut content = String::from("# Coven Chat Export\n\n");
+        for msg in &self.messages {
+            let role_label = match msg.role {
+                MessageRole::User => "**You**",
+                MessageRole::Agent => &format!("**{}**", msg.sender),
+                MessageRole::System => "*system*",
+            };
+            content.push_str(&format!(
+                "{} ({})\n{}\n\n---\n\n",
+                role_label, msg.timestamp, msg.content
+            ));
+        }
+
+        match std::fs::write(&path, content) {
+            Ok(()) => self.push_system_message(&format!("Exported to {}", path.display())),
+            Err(e) => self.push_system_message(&format!("Export failed: {e}")),
+        }
+    }
+
+    pub(super) fn scroll_to_bottom(&mut self) {
+        // Will be calculated during render based on content height
+        self.scroll_offset = usize::MAX;
+    }
+
+    pub(super) fn tick(&mut self) {
+        if self.last_tick.elapsed() >= Duration::from_millis(120) {
+            self.spinner_frame = (self.spinner_frame + 1) % SPINNER_FRAMES.len();
+            self.last_tick = Instant::now();
+        }
+    }
+
+    pub(super) fn insert_char(&mut self, c: char) {
+        self.input.insert(self.cursor_pos, c);
+        self.cursor_pos += c.len_utf8();
+    }
+
+    pub(super) fn delete_char_before_cursor(&mut self) {
+        if self.cursor_pos > 0 {
+            let prev = self.input[..self.cursor_pos]
+                .chars()
+                .last()
+                .map(|c| c.len_utf8())
+                .unwrap_or(0);
+            self.cursor_pos -= prev;
+            self.input.remove(self.cursor_pos);
+        }
+    }
+
+    pub(super) fn delete_char_at_cursor(&mut self) {
+        if self.cursor_pos < self.input.len() {
+            self.input.remove(self.cursor_pos);
+        }
+    }
+
+    pub(super) fn move_cursor_left(&mut self) {
+        if self.cursor_pos > 0 {
+            let prev = self.input[..self.cursor_pos]
+                .chars()
+                .last()
+                .map(|c| c.len_utf8())
+                .unwrap_or(0);
+            self.cursor_pos -= prev;
+        }
+    }
+
+    pub(super) fn move_cursor_right(&mut self) {
+        if self.cursor_pos < self.input.len() {
+            let next = self.input[self.cursor_pos..]
+                .chars()
+                .next()
+                .map(|c| c.len_utf8())
+                .unwrap_or(0);
+            self.cursor_pos += next;
+        }
+    }
+
+    pub(super) fn move_cursor_home(&mut self) {
+        self.cursor_pos = 0;
+    }
+
+    pub(super) fn move_cursor_end(&mut self) {
+        self.cursor_pos = self.input.len();
+    }
+
+    pub(super) fn delete_word_before_cursor(&mut self) {
+        if self.cursor_pos == 0 {
+            return;
+        }
+        let before = &self.input[..self.cursor_pos];
+        let trimmed = before.trim_end();
+        let new_end = trimmed
+            .rfind(char::is_whitespace)
+            .map(|i| i + 1)
+            .unwrap_or(0);
+        self.input.drain(new_end..self.cursor_pos);
+        self.cursor_pos = new_end;
+    }
+}
+
+// ── Discover agents from built-in harnesses ────────────────────────────────
+
+pub(super) fn discover_agents() -> Vec<AgentInfo> {
+    harness::built_in_harnesses()
+        .into_iter()
+        .map(|h| AgentInfo {
+            id: h.id.to_string(),
+            label: h.label.to_string(),
+            harness: h.id.to_string(),
+            available: h.available,
+        })
+        .collect()
+}
+
+// ── Helpers ────────────────────────────────────────────────────────────────
+
+fn timestamp_now() -> String {
+    chrono::Local::now().format("%H:%M").to_string()
+}
+
+fn truncate_str(s: &str, max: usize) -> &str {
+    if s.len() <= max {
+        s
+    } else {
+        let end = s.char_indices().nth(max).map(|(i, _)| i).unwrap_or(s.len());
+        &s[..end]
+    }
+}
+
+#[cfg(test)]
+mod tests {
+    use super::*;
+
+    fn app_with_agents(agents: Vec<AgentInfo>) -> App {
+        let active_agent = agents.iter().position(|agent| agent.available);
+        App {
+            messages: Vec::new(),
+            input: String::new(),
+            cursor_pos: 0,
+            scroll_offset: 0,
+            agents,
+            active_agent,
+            input_mode: InputMode::Normal,
+            agent_select_index: 0,
+            show_help: false,
+            spinner_frame: 0,
+            is_responding: false,
+            last_tick: Instant::now(),
+        }
+    }
+
+    fn agent(id: &str, available: bool) -> AgentInfo {
+        AgentInfo {
+            id: id.to_string(),
+            label: id.to_string(),
+            harness: id.to_string(),
+            available,
+        }
+    }
+
+    #[test]
+    fn unknown_slash_command_returns_command_name_for_feedback() {
+        let mut app = app_with_agents(vec![agent("codex", true)]);
+
+        match app.handle_slash_command("/unknown value") {
+            SlashCommandResult::Unknown(command) => assert_eq!(command, "/unknown"),
+            other => panic!("expected unknown command result, got {other:?}"),
+        }
+    }
+
+    #[test]
+    fn handle_input_clears_unknown_slash_command_and_reports_it() {
+        let mut app = app_with_agents(vec![agent("codex", true)]);
+        app.input = "/missing".to_string();
+        app.cursor_pos = app.input.len();
+
+        let result = app.handle_input();
+
+        match result {
+            Some(SlashCommandResult::Unknown(command)) => assert_eq!(command, "/missing"),
+            other => panic!("expected unknown command result, got {other:?}"),
+        }
+        assert!(app.input.is_empty());
+        assert_eq!(app.cursor_pos, 0);
+    }
+
+    #[test]
+    fn agent_command_without_argument_opens_picker_on_active_agent() {
+        let mut app = app_with_agents(vec![agent("claude", false), agent("codex", true)]);
+
+        let result = app.handle_slash_command("/agent");
+
+        assert!(matches!(result, SlashCommandResult::Handled));
+        assert_eq!(app.input_mode, InputMode::AgentSelect);
+        assert_eq!(app.agent_select_index, 1);
+    }
+
+    #[test]
+    fn unavailable_agent_selection_keeps_current_active_agent() {
+        let mut app = app_with_agents(vec![agent("claude", false), agent("codex", true)]);
+
+        app.switch_agent_by_name("claude");
+
+        assert_eq!(app.active_agent, Some(1));
+        assert!(app
+            .messages
+            .last()
+            .map(|message| message.content.contains("claude is not available"))
+            .unwrap_or(false));
+    }
+}
