+++ crates/coven-cli/src/tui/mod.rs
@@ -1,0 +1,4 @@
+//! TUI surfaces for the coven CLI. Currently hosts the chat module; Phases 3–4
+//! will land the launcher and session-browser carve-outs from main.rs here.
+
+pub mod chat;
