+++ scripts/check-secrets.py
@@ -79,6 +79,23 @@ def is_known_safe_lockfile_line(path: str, line: str) -> bool:
     )
 
 
+def is_local_path_like_token(token: str) -> bool:
+    normalized = token.strip("/")
+    parts = normalized.split("/")
+    if len(parts) < 4:
+        return False
+    if parts[0] in {"Users", "home", "private", "var", "tmp", "Volumes"}:
+        return True
+    return ".worktrees" in parts or "worktrees" in parts
+
+
+def is_public_repo_url_like_token(token: str) -> bool:
+    normalized = token.strip("/")
+    return normalized.startswith("github.com/OpenCoven/coven/") and (
+        "/blob/" in normalized or "/tree/" in normalized
+    )
+
+
 def scan_text(text: str, path: str) -> list[tuple[str, int, str]]:
     hits: list[tuple[str, int, str]] = []
     for line_number, line in enumerate(text.splitlines(), 1):
@@ -95,6 +112,8 @@ def scan_text(text: str, path: str) -> list[tuple[str, int, str]]:
         for match in re.finditer(r"\b[A-Za-z0-9_+/@.-]{32,}\b", line):
             token = match.group(0)
             if re.fullmatch(r"[0-9a-f]{32,64}", token):
+                continue
+            if is_local_path_like_token(token) or is_public_repo_url_like_token(token):
                 continue
             if entropy(token) >= 4.3:
                 hits.append((path, line_number, "high_entropy"))
